feat(xras): a stress tier whose assertion target is the audit row
## Summary

`tests/stress/`, gated behind a new `-m stress` marker mirroring the perf tier, with
a `scenarios.json` manifest mirroring `baselines.json`. 17 scenarios.

The organising principle is not "stress the handlers". It is **stress the handlers
and assert on the audit row** — because cutover is an abrupt repoint with no
observation window, and that row is the only artefact that survives into triage
week. Where a scenario cannot be triaged from the row alone, it says so in its
`verdict` field and stays green: schema evidence, not a failing test.

Scenarios run through the **HTTP route**, not `dispatch_action`, since the row is
what is under test. Two consequences are load-bearing and documented in the fixtures:
route tests cannot use factories (the route reads `db.session`, which sees only
committed rows — a factory project silently turns a scenario into "no service
matched"), and the rows the route commits need explicit cleanup, which is why
`action_log` moved to `tests/xras_audit.py` and is now shared rather than copied.

## The audit write could be lost, and that is measured

`raw_payload` and `error_messages` were written unbounded into `TEXT`. Under
`STRICT_TRANS_TABLES` — confirmed on — an oversized value does not truncate, it
raises `1406 Data too long`. Reproduced live before the guard existed: the INSERT
failed and the row was gone. That is the one failure this table cannot afford, and
`_fit()` already existed on two neighbouring columns for exactly this reason.

**Which path can reach it took the corrected wire field name to see**, and the
answer moved once `resourceRepositoryKey` replaced `key`:

| path | ratio | reachable? |
|---|---|---|
| Supplement | **1.00x** | no — a failed key `continue`s before the amount is read, so one message per resource against an entry of near-identical length. The body hits its limit first |
| New | **1.79x** | **yes** — `_plan_allocations` calls `resolve_resource` *and* `transaction_amount` unconditionally, so one resource yields two messages. 59,090 bytes of body → 105,999 of messages |

So the guard sits on the only path that reaches the condition. Both measurements are
in the manifest, and the contrast is its own test so it cannot silently drift.

Truncation is self-announcing in both columns, cuts `error_messages` on **message**
boundaries (half a diagnostic reads as a whole one), and never touches the 422 body —
that is the wire contract XRAS admins read. An oversized body is **refused** rather
than applied, because a body we cannot record is one we cannot audit or replay.

## Four parks, one row, no way to tell them apart

`test_a_disabled_park_and_an_unmatched_park_are_byte_identical` asserts the equality
directly over the six columns an operator can filter on.

⚠️ That test passing **is** the problem. It must go red the day a `service` /
`outcome_reason` column lands, at which point it becomes the test that proves the
column works.

## Three questions the corpus could not answer, now answered

- **`Co-PI` vs `CoPi`** — the spelling does not matter. Membership takes every
  username in `roles[]` regardless of `roleType`; assignment matches only `PI` and
  `Allocation Manager`, so an unknown spelling is simply not a candidate. Tested
  across three spellings. This retires it as a cutover risk and the unanswered
  `hdt@ucar.edu` sample with it.
- **`Renewal`** — reaches `UpdateHandler` and returns six real error strings. A
  minimal payload failing validation is the *correct* outcome; the claim is that an
  unsampled type produces a reviewable 422 rather than legacy's opaque 500.
- **`Advance`** — parks, as designed, with no service to match.

And the double-post blast radius, measured at the level where it is observable:
Supplement is additive (250,000 three times leaves 750,000 added), Extension is
near-idempotent via the equal-end-date skip. Extension is 60% of traffic and costs
nothing; Supplement is 15% and costs a full increment. That asymmetry is the evidence
behind the `action_id` verdict.

## Also

- `http_status` → `SmallInteger`, matching `SMALLINT UNSIGNED`, pinned by a test that
  compares every declared width against the live DDL.
- `processed_by`'s width slice moved from four `replay.py` call sites into `_record`,
  where the column is.

### Test Results

5,233 passed (17 new stress scenarios run separately via `pytest -m stress`).
Stress tier verified xdist-safe. Leak check → 0 across `allocation_transaction`,
`xras_action_log` and `project`.

+++ src/sam/integration/xras.py
@@ -128,11 +128,15 @@ class XrasActionLog(Base):
     #: received | processed | manual | failed | replayed
     status = Column(String(16), nullable=False)
 
-    #: The HTTP code we answered: 200, 400 or 422. ``status='failed'`` covers both a
-    #: malformed body (400) and a schema rejection (422) — an operator triaging the
-    #: log needs to tell those apart, and it stops being derivable from ``status``
-    #: the moment handlers add their own validation failures.
-    http_status = Column(Integer)
+    #: The HTTP code we answered: 200, 400 or 422. ``status='failed'`` covers a
+    #: malformed body (400), a schema rejection (422), a handler rejection (422) and
+    #: an oversized body (422) — an operator triaging the log needs to tell those
+    #: apart, and it stops being derivable from ``status``.
+    #:
+    #: ``SmallInteger`` to match the DDL's ``SMALLINT UNSIGNED``. It was ``Integer``,
+    #: which is harmless in MySQL but is the kind of drift that makes a guard computed
+    #: from the ORM quietly wrong — pinned by ``tests/stress/test_audit_row_survives.py``.
+    http_status = Column(SmallInteger)
 
     #: The ordered error list, one message per line — the same list the 422 carries.
     error_messages = Column(Text)
